Move volatility function as helper in cap level endpoint
