+++ crates/server/src/pricing_data/mod.rs
@@ -2,4 +2,3 @@ pub mod cap_level;
 pub mod reserve_price;
 pub mod twap;
 mod utils;
-pub mod volatility;
